Extract static function as `x509_get_pka()`
